+++ packages/login-panel/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@3.1.2...@rtm-ui/login-panel@3.1.3) (2019-01-30)
+
+
+### Bug Fixes
+
+* **login-panel:** fix icon display with obs and fiftyup ([464763e](https://github.com/OBSAU/rtm-kit/commit/464763e))
+* **login-panel:** reafactor code display color for icon ([2a23f94](https://github.com/OBSAU/rtm-kit/commit/2a23f94))
+* **login-panel:** update implement use button icon ([e4b5d04](https://github.com/OBSAU/rtm-kit/commit/e4b5d04))
+
+
+
+
+
 ## [3.1.2](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@3.1.1...@rtm-ui/login-panel@3.1.2) (2019-01-25)
 
 **Note:** Version bump only for package @rtm-ui/login-panel